Work around compiler error
error: reinterpret_cast from 'Pixmap' (aka 'unsigned long') to 'EGLNativePixmapType' (aka 'unsigned __intcap') is not allowed

+++ src/hardwareintegration/compositor/xcomposite-egl/xcompositeeglintegration.cpp
@@ -116,7 +116,7 @@ QOpenGLTexture *XCompositeEglClientBuffer::toOpenGlTexture(int plane)
     attribList.append(EGL_TEXTURE_2D);
     attribList.append(EGL_NONE);
 
-    EGLSurface surface = eglCreatePixmapSurface(m_integration->eglDisplay(), config, reinterpret_cast<EGLNativePixmapType>(pixmap), attribList.constData());
+    EGLSurface surface = eglCreatePixmapSurface(m_integration->eglDisplay(), config, reinterpret_cast<EGLNativePixmapType>((uintptr_t)pixmap), attribList.constData());
     if (surface == EGL_NO_SURFACE) {
         qDebug() << "Failed to create eglsurface" << pixmap << compositorBuffer->window();
     }
